Move pieces out of profile.js
