+++ src/lib/utils_misc.js
@@ -172,6 +172,12 @@ const utilsMisc = {
 
   },
 
+  // set string character at index to chr
+  setCharAt: function (str, index, chr) {
+      if (index > str.length - 1) return str;
+      return str.substring(0, index) + chr + str.substring(index + 1);
+  },
+
 
 
 }
